nuget pack: Add -p as an explicit alias for -Properties
Previously, the only parameter to `nuget pack` that started with 'p' was
'Properties'.  A number of unit tests relied on this, and abbreviated it
as -p.  The introduction of -PackagesDir caused these unit tests to fail
because '-p' became ambiguous.

This commit adds '-p' as an explicit alias for '-Properties'; this way,
scripts that used -p will continue to work unmodified.